Address further review feedback: Revert unrelated changes, move CDI producers, and fix schema bootstrap for isolated DataSources
