+++ persistence/relational-jdbc/src/main/java/org/apache/polaris/persistence/relational/jdbc/RelationalJdbcConfiguration.java
@@ -36,9 +36,4 @@ public interface RelationalJdbcConfiguration {
    */
   Optional<String> databaseType();
 
-  /**
-   * The identifier of the {@link DataSourceResolver} bean to use. Used by {@link JdbcCdiProducers}
-   * to select the resolver via CDI's {@link io.smallrye.common.annotation.Identifier}.
-   */
-  String dataSourceResolverType();
 }
