+++ persistence/relational-jdbc/src/test/java/org/apache/polaris/persistence/relational/jdbc/AtomicMetastoreManagerWithJdbcBasePersistenceImplTest.java
@@ -103,10 +103,5 @@ public abstract class AtomicMetastoreManagerWithJdbcBasePersistenceImplTest
     public Optional<String> databaseType() {
       return Optional.of("h2");
     }
-
-    @Override
-    public String dataSourceResolverType() {
-      return "default";
-    }
   }
 }
