+++ persistence/relational-jdbc/src/test/java/org/apache/polaris/persistence/relational/jdbc/MetricsReportPersistenceTest.java
@@ -204,10 +204,5 @@ class MetricsReportPersistenceTest {
     public Optional<String> databaseType() {
       return Optional.empty();
     }
-
-    @Override
-    public String dataSourceResolverType() {
-      return "default";
-    }
   }
 }
