+++ persistence/relational-jdbc/src/test/java/org/apache/polaris/persistence/relational/jdbc/idempotency/RelationalJdbcIdempotencyStorePostgresIT.java
@@ -80,12 +80,7 @@ public class RelationalJdbcIdempotencyStorePostgresIT {
 
           @Override
           public Optional<String> databaseType() {
-            return Optional.empty();
-          }
-
-          @Override
-          public String dataSourceResolverType() {
-            return "default";
+            return Optional.of("postgresql");
           }
         };
     DatasourceOperations ops = new DatasourceOperations(dataSource, cfg);
